Extract shared panel components and simplify mobile experience
Extracted reusable sub-components from HighlightsPanel into shared-panel-components.tsx
for better separation. Hid v1/v2 compare toggles on mobile since the panel isn't
side-by-side. Removed unused onStoryTap prop from DecisionStories.
