+++ src/components/vignettes/ai-highlights/HighlightsPanel.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useState, useEffect, useRef, useCallback } from 'react';
+import { useState, useEffect, useRef } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 import { useReducedMotion } from '@/lib/useReducedMotion';
 import { useIntroSequence } from '@/lib/intro-sequence-context';
@@ -8,154 +8,25 @@ import { DecisionStory } from '../shared/DecisionStories';
 import {
   aiHighlightsContent,
   storyInteractions,
-  FeedbackSource,
   HighlightItem,
-  BeforeSummaryItem,
-  SummaryPart,
 } from './content';
+import {
+  WIREFRAME_GRAY,
+  getAvatarColor,
+  getInitials,
+  SourceCard,
+  BeforeSummaryItemRow,
+  DetailSpan,
+  FocusedItem,
+  CollapsibleItem,
+  LoadingStyles,
+  FloatingThumbs,
+} from './shared-panel-components';
 
 interface HighlightsPanelProps {
   className?: string;
   activeStory?: DecisionStory | null;
   showBeforeState?: boolean;
-}
-
-interface SourceCardProps {
-  source: FeedbackSource;
-}
-
-function SourceCard({ source }: SourceCardProps) {
-  return (
-    <div className="pt-4 border-t border-border">
-      {/* Mobile: stacked layout */}
-      <div className="flex flex-col gap-1 mb-3 sm:hidden">
-        <div className="flex items-center gap-2">
-          <div
-            className="w-6 h-6 rounded-full flex items-center justify-center text-[10px] font-bold text-white"
-            style={{ backgroundColor: getAvatarColor(source.reviewer) }}
-          >
-            {getInitials(source.reviewer)}
-          </div>
-          <span className="text-body-sm font-semibold text-primary">
-            {source.reviewer}
-          </span>
-        </div>
-        <div className="flex items-center gap-2 text-caption text-secondary">
-          <span>{source.reviewerRole}</span>
-        </div>
-      </div>
-      {/* Desktop: horizontal layout */}
-      <div className="hidden sm:flex items-center gap-2 mb-3">
-        <div
-          className="w-6 h-6 rounded-full flex items-center justify-center text-[10px] font-bold text-white"
-          style={{ backgroundColor: getAvatarColor(source.reviewer) }}
-        >
-          {getInitials(source.reviewer)}
-        </div>
-        <span className="text-body-sm font-semibold text-primary">
-          {source.reviewer}
-        </span>
-        <span className="text-caption text-secondary">•</span>
-        <span className="text-caption text-secondary">
-          {source.reviewerRole}
-        </span>
-      </div>
-      <p className="text-body-sm text-primary mb-3">{source.quote}</p>
-    </div>
-  );
-}
-
-// Helper functions for avatar colors
-function getInitials(name: string): string {
-  return name
-    .split(' ')
-    .map((n) => n[0])
-    .join('')
-    .toUpperCase();
-}
-
-function getAvatarColor(name: string): string {
-  const colors: Record<string, string> = {
-    'David Park': '#6366F1',
-    'Rachel Torres': '#EC4899',
-    'James Liu': '#10B981',
-    'Priya Sharma': '#F59E0B',
-    'Kevin Wright': '#8B5CF6',
-  };
-  return colors[name] || '#6366F1';
-}
-
-function LoadingStyles() {
-  return (
-    <style jsx global>{`
-      @property --gradient-angle {
-        syntax: '<angle>';
-        initial-value: 0deg;
-        inherits: false;
-      }
-
-      @keyframes rotateGradient {
-        to {
-          --gradient-angle: 360deg;
-        }
-      }
-
-      .loading-panel-border {
-        position: absolute;
-        inset: 0;
-        border-radius: 7px;
-        background: conic-gradient(
-          from var(--gradient-angle),
-          var(--ai-gradient-1),
-          var(--ai-gradient-2),
-          var(--ai-gradient-3),
-          var(--ai-gradient-2),
-          var(--ai-gradient-1)
-        );
-        animation: rotateGradient 3s linear infinite;
-        filter: drop-shadow(0 0 20px rgba(166, 229, 231, 0.5));
-      }
-
-      .loading-panel-content {
-        position: relative;
-        background: var(--background-elevated);
-        width: 100%;
-        height: 100%;
-        border-radius: 5px;
-        z-index: 1;
-      }
-
-      @keyframes shimmer {
-        0% {
-          background-position: -200% 0;
-        }
-        100% {
-          background-position: 200% 0;
-        }
-      }
-
-      .skeleton-bar {
-        background: linear-gradient(
-          90deg,
-          rgba(0, 0, 0, 0.06) 25%,
-          rgba(0, 0, 0, 0.1) 50%,
-          rgba(0, 0, 0, 0.06) 75%
-        );
-        background-size: 200% 100%;
-        animation: shimmer 1.5s ease-in-out infinite;
-      }
-
-      @media (prefers-reduced-motion: reduce) {
-        .loading-panel-border {
-          animation: none;
-          filter: drop-shadow(0 0 20px rgba(166, 229, 231, 0.3));
-        }
-        .skeleton-bar {
-          animation: none;
-        }
-      }
-    `}</style>
-  );
 }
 
 function SkeletonContent() {
@@ -215,454 +86,10 @@ function LoadingState() {
   );
 }
 
-const WIREFRAME_GRAY = '#C4C4C4';
-
-function BeforeSummaryItemRow({ item }: { item: BeforeSummaryItem }) {
-  return (
-    <div className="border-b-2 border-border">
-      <div className="px-6 py-8">
-        <div className="flex flex-col sm:flex-row sm:items-center gap-3">
-          <div className="flex-1">
-            <p className="text-body-sm text-primary">{item.summary}</p>
-          </div>
-
-          <div className="flex items-center gap-3 shrink-0 px-2 py-2 -mx-2">
-            <div className="flex items-center gap-1">
-              <div className="flex -space-x-2">
-                {item.sources.slice(0, 2).map((source, idx) => (
-                  <div
-                    key={idx}
-                    className="w-5 h-5 rounded-full border-2 border-background-elevated flex items-center justify-center text-[8px] font-bold text-white"
-                    style={{ backgroundColor: WIREFRAME_GRAY }}
-                  >
-                    {getInitials(source.reviewer)}
-                  </div>
-                ))}
-              </div>
-              <span className="text-body-sm text-secondary">
-                {item.sources.length} sources
-              </span>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function DetailSpan({
-  detail,
-  animateExpand,
-  delayIndex,
-}: {
-  detail: string;
-  animateExpand: boolean;
-  delayIndex: number;
-}) {
-  const highlightDelay = 0.1 + delayIndex * 0.15;
-  return (
-    <motion.span
-      className="rounded-sm"
-      style={{ display: 'inline', boxDecorationBreak: 'clone' }}
-      initial={
-        animateExpand
-          ? { backgroundColor: 'rgba(253, 224, 71, 0.45)' }
-          : false
-      }
-      animate={{ backgroundColor: 'rgba(253, 224, 71, 0)' }}
-      transition={{
-        backgroundColor: {
-          duration: 1.4,
-          delay: animateExpand ? highlightDelay + 0.4 : 0,
-          ease: 'easeOut',
-        },
-      }}
-    >
-      {detail}
-    </motion.span>
-  );
-}
-
-function FocusedItem({
-  item,
-  summaryParts,
-  animateExpand,
-}: {
-  item: HighlightItem;
-  summaryParts?: SummaryPart[];
-  animateExpand: boolean;
-}) {
-  const isHighlight = item.type === 'highlight';
-  const label = isHighlight ? 'Highlight' : 'Opportunity';
-  const icon = isHighlight ? 'star_outline' : 'lightbulb';
-  const iconColor = isHighlight ? '#22594A' : 'rgba(135, 100, 0, 1)';
-
-  const renderSummary = () => {
-    if (!summaryParts || !animateExpand) {
-      return (
-        <p className="text-body-sm text-primary">{item.summary}</p>
-      );
-    }
-
-    let detailIndex = 0;
-    return (
-      <p className="text-body-sm text-primary">
-        {summaryParts.map((part, i) => {
-          if (typeof part === 'string') {
-            return <span key={i}>{part}</span>;
-          }
-          const idx = detailIndex++;
-          return (
-            <DetailSpan
-              key={i}
-              detail={part.detail}
-              animateExpand={animateExpand}
-              delayIndex={idx}
-            />
-          );
-        })}
-      </p>
-    );
-  };
-
-  return (
-    <div className="border-b-2 border-border">
-      <div className="px-6 py-6">
-        <div className="flex flex-col sm:flex-row sm:items-center gap-3">
-          <div className="flex-1">
-            <div className="flex items-center gap-1 mb-2">
-              <span
-                className="material-icons-outlined text-h3"
-                style={{ color: iconColor }}
-              >
-                {icon}
-              </span>
-              <span className="text-body-sm font-semibold text-primary">
-                {label}
-              </span>
-            </div>
-            {renderSummary()}
-          </div>
-
-          <div className="flex items-center gap-3 shrink-0 px-2 py-2 -mx-2 opacity-30">
-            <div className="flex items-center gap-1">
-              <div className="flex -space-x-2">
-                {item.sources.slice(0, 2).map((source, idx) => (
-                  <div
-                    key={idx}
-                    className="w-5 h-5 rounded-full border-2 border-background-elevated flex items-center justify-center text-[8px] font-bold text-white"
-                    style={{ backgroundColor: getAvatarColor(source.reviewer) }}
-                  >
-                    {getInitials(source.reviewer)}
-                  </div>
-                ))}
-              </div>
-              <span className="text-body-sm text-secondary">
-                {item.sources.length} sources
-              </span>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-
-// --- Floating Thumbs Animation ---
-
-interface FloatingThumb {
-  id: number;
-  /** Horizontal offset from center in px */
-  startX: number;
-  /** Horizontal drift destination in px */
-  driftX: number;
-  /** Vertical rise distance in px */
-  riseDistance: number;
-  /** Total animation duration in seconds */
-  duration: number;
-  /** Font size in px */
-  size: number;
-  /** Fixed rotation in degrees */
-  rotation: number;
-  /** Starting scale */
-  startScale: number;
-  /** Ending scale */
-  endScale: number;
-}
-
-let thumbIdCounter = 0;
-
-function createFloatingThumb(): FloatingThumb {
-  const id = thumbIdCounter++;
-  // Spawn within ~40px horizontal band around center
-  const startX = (Math.random() - 0.5) * 40;
-  // Horizontal drift: 20-50px amplitude, alternating direction
-  const driftDirection = Math.random() > 0.5 ? 1 : -1;
-  const driftX = driftDirection * (20 + Math.random() * 30);
-  // Vertical rise: 120-180px
-  const riseDistance = 120 + Math.random() * 60;
-  // Duration: 2.5-3.5s
-  const duration = 2.5 + Math.random() * 1.0;
-  // Size: 14-18px
-  const size = 14 + Math.random() * 4;
-  // Fixed rotation: -25 to +25 degrees
-  const rotation = (Math.random() - 0.5) * 50;
-  // Scale: start 0.6-0.9, end 0.6-0.8
-  const startScale = 0.6 + Math.random() * 0.3;
-  const endScale = 0.6 + Math.random() * 0.2;
-
-  return {
-    id,
-    startX,
-    driftX,
-    riseDistance,
-    duration,
-    size,
-    rotation,
-    startScale,
-    endScale,
-  };
-}
-
-function FloatingThumbParticle({
-  thumb,
-  onComplete,
-}: {
-  thumb: FloatingThumb;
-  onComplete: (id: number) => void;
-}) {
-  return (
-    <motion.div
-      style={{
-        position: 'absolute',
-        left: '50%',
-        bottom: 0,
-        pointerEvents: 'none',
-        fontSize: thumb.size,
-        lineHeight: 1,
-        willChange: 'transform, opacity',
-        rotate: thumb.rotation,
-      }}
-      initial={{
-        x: thumb.startX,
-        y: 0,
-        opacity: 0,
-        scale: thumb.startScale,
-      }}
-      animate={{
-        x: thumb.startX + thumb.driftX,
-        y: -thumb.riseDistance,
-        opacity: [0, 1, 1, 0],
-        scale: [thumb.startScale, 1, thumb.endScale + 0.1, thumb.endScale],
-      }}
-      transition={{
-        y: {
-          duration: thumb.duration,
-          ease: [0.2, 0.6, 0.4, 1],
-        },
-        x: {
-          duration: thumb.duration,
-          ease: 'easeInOut',
-        },
-        opacity: {
-          duration: thumb.duration,
-          times: [0, 0.1, 0.7, 1],
-          ease: 'linear',
-        },
-        scale: {
-          duration: thumb.duration,
-          times: [0, 0.15, 0.7, 1],
-          ease: 'easeOut',
-        },
-      }}
-      onAnimationComplete={() => onComplete(thumb.id)}
-    >
-      <span className="material-icons-outlined text-emerald-400" style={{ fontSize: 'inherit' }}>
-        thumb_up
-      </span>
-    </motion.div>
-  );
-}
-
-function FloatingThumbs({ active, burstTrigger = 0 }: { active: boolean; burstTrigger?: number }) {
-  const [particles, setParticles] = useState<FloatingThumb[]>([]);
-  const reducedMotion = useReducedMotion();
-  const intervalRef = useRef<ReturnType<typeof setInterval> | null>(null);
-
-  const removeParticle = useCallback((id: number) => {
-    setParticles((prev) => prev.filter((p) => p.id !== id));
-  }, []);
-
-  useEffect(() => {
-    if (!active || reducedMotion) {
-      if (intervalRef.current) {
-        clearInterval(intervalRef.current);
-        intervalRef.current = null;
-      }
-      setParticles([]);
-      return;
-    }
-
-    // Spawn first particle immediately
-    setParticles([createFloatingThumb()]);
-
-    // Then spawn every 350ms
-    intervalRef.current = setInterval(() => {
-      setParticles((prev) => {
-        if (prev.length >= 10) return prev;
-        return [...prev, createFloatingThumb()];
-      });
-    }, 350);
-
-    return () => {
-      if (intervalRef.current) {
-        clearInterval(intervalRef.current);
-        intervalRef.current = null;
-      }
-    };
-  }, [active, reducedMotion]);
-
-  // Burst: spawn 3-4 extra particles on click
-  useEffect(() => {
-    if (burstTrigger === 0 || !active || reducedMotion) return;
-    const burst = Array.from({ length: 3 + Math.floor(Math.random() * 2) }, () => createFloatingThumb());
-    setParticles((prev) => [...prev, ...burst]);
-  }, [burstTrigger, active, reducedMotion]);
-
-  if (reducedMotion || !active) return null;
-
-  return (
-    <div
-      style={{
-        position: 'absolute',
-        left: '50%',
-        bottom: '100%',
-        width: 0,
-        height: 0,
-        pointerEvents: 'none',
-        zIndex: 10,
-        overflow: 'visible',
-      }}
-    >
-      <AnimatePresence>
-        {particles.map((thumb) => (
-          <FloatingThumbParticle
-            key={thumb.id}
-            thumb={thumb}
-            onComplete={removeParticle}
-          />
-        ))}
-      </AnimatePresence>
-    </div>
-  );
-}
-
-// --- End Floating Thumbs Animation ---
-
 interface SolutionStateProps {
   className?: string;
   activeStory?: DecisionStory | null;
   showBeforeState?: boolean;
-}
-
-function CollapsibleItem({
-  item,
-  sectionIndex,
-  isExpanded,
-  onToggle,
-  blockStyle,
-}: {
-  item: HighlightItem;
-  sectionIndex: number;
-  isExpanded: boolean;
-  onToggle: () => void;
-  blockStyle: React.CSSProperties;
-}) {
-  const isHighlight = item.type === 'highlight';
-  const label = isHighlight ? 'Highlight' : 'Opportunity';
-  const icon = isHighlight ? 'star_outline' : 'lightbulb';
-  const iconColor = isHighlight ? '#22594A' : 'rgba(135, 100, 0, 1)';
-
-  return (
-    <div
-      className="border-b-2 border-border"
-      data-section-id={`${item.type}-${sectionIndex}`}
-      style={blockStyle}
-    >
-      <div className="px-6 py-8">
-        <div className="flex flex-col sm:flex-row sm:items-center gap-3">
-          <div className="flex-1">
-            <div className="flex items-center gap-1 mb-2">
-              <span
-                className="material-icons-outlined text-h3"
-                style={{ color: iconColor }}
-              >
-                {icon}
-              </span>
-              <span className="text-body-sm font-semibold text-primary">
-                {label}
-              </span>
-            </div>
-            <p className="text-body-sm text-primary">
-              {item.summary}
-            </p>
-          </div>
-
-          <button
-            onClick={onToggle}
-            className="flex items-center gap-3 shrink-0 hover:bg-black/5 rounded-lg px-2 py-2 -mx-2 transition-colors cursor-pointer"
-            aria-label={
-              isExpanded ? `Collapse ${item.type}` : `Expand ${item.type}`
-            }
-            aria-expanded={isExpanded}
-          >
-            <div className="flex items-center gap-1">
-              <div className="flex -space-x-2">
-                {item.sources.slice(0, 2).map((source, idx) => (
-                  <div
-                    key={idx}
-                    className="w-5 h-5 rounded-full border-2 border-background-elevated flex items-center justify-center text-[8px] font-bold text-white"
-                    style={{ backgroundColor: getAvatarColor(source.reviewer) }}
-                  >
-                    {getInitials(source.reviewer)}
-                  </div>
-                ))}
-              </div>
-              <span className="text-body-sm text-secondary">
-                {item.sources.length} sources
-              </span>
-            </div>
-            <motion.span
-              animate={{ rotate: isExpanded ? 180 : 0 }}
-              transition={{ duration: 0.3 }}
-              className="material-icons-outlined text-h3 text-primary block"
-            >
-              keyboard_arrow_down
-            </motion.span>
-          </button>
-        </div>
-
-        <AnimatePresence>
-          {isExpanded && (
-            <motion.div
-              initial={{ height: 0, opacity: 0 }}
-              animate={{ height: 'auto', opacity: 1 }}
-              exit={{ height: 0, opacity: 0 }}
-              transition={{ duration: 0.3, ease: 'easeInOut' }}
-              className="overflow-hidden"
-            >
-              <div className="mt-4 space-y-4">
-                {item.sources.map((source, idx) => (
-                  <SourceCard key={idx} source={source} />
-                ))}
-              </div>
-            </motion.div>
-          )}
-        </AnimatePresence>
-      </div>
-    </div>
-  );
 }
 
 function SolutionState({
@@ -677,8 +104,6 @@ function SolutionState({
   const [showStat, setShowStat] = useState(false);
   const [borderAnimating, setBorderAnimating] = useState(false);
   const [burstTrigger, setBurstTrigger] = useState(0);
-  // 0 = all v1, totalSections = all v2
-  // Sections: 0=header, 1..N=items
   const totalSections = allItems.length + 1;
   const [upgradedSections, setUpgradedSections] = useState(0);
   const prevBeforeStateRef = useRef(showBeforeState);
@@ -691,13 +116,11 @@ function SolutionState({
   const isMeasuringStory = activeStory?.id === 'measuring-success';
   const hasToggle = isReframingStory || isRefiningStory;
 
-  // Reset per-story state when activeStory changes
   useEffect(() => {
     setThumbsClicked(false);
     setShowStat(false);
     setBorderAnimating(false);
     setBurstTrigger(0);
-    // When a new story opens, show the correct initial state instantly
     setUpgradedSections(showBeforeState ? 0 : totalSections);
     prevBeforeStateRef.current = showBeforeState;
 
@@ -717,7 +140,6 @@ function SolutionState({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [activeStory]);
 
-  // Staggered v1 → v2 streaming upgrade
   useEffect(() => {
     const isToggleStory = isReframingStory || isRefiningStory;
     const wasBeforeNowAfter =
@@ -725,7 +147,6 @@ function SolutionState({
 
     prevBeforeStateRef.current = showBeforeState;
 
-    // v2 → v1: instant reset
     if (showBeforeState) {
       setUpgradedSections(0);
       setBorderAnimating(false);
@@ -738,17 +159,15 @@ function SolutionState({
         return;
       }
 
-      // 1. Border starts glowing
       setBorderAnimating(true);
 
       const timers: ReturnType<typeof setTimeout>[] = [];
 
       if (isRefiningStory) {
-        // Header doesn't visually change — upgrade it immediately
         setUpgradedSections(1);
         const itemDelay = 400;
         const itemInterval = 900;
-        const itemCount = totalSections - 1; // exclude header
+        const itemCount = totalSections - 1;
         for (let i = 0; i < itemCount; i++) {
           timers.push(
             setTimeout(
@@ -757,16 +176,13 @@ function SolutionState({
             )
           );
         }
-        // Settle after last expand animation finishes
         timers.push(
           setTimeout(() => setBorderAnimating(false),
             itemDelay + (itemCount - 1) * itemInterval + 1500
           )
         );
       } else {
-        // Reframing: upgrade all sections at once for a simple crossfade
         setUpgradedSections(totalSections);
-        // Brief border glow flash, then stop
         timers.push(
           setTimeout(() => setBorderAnimating(false), 1200)
         );
@@ -776,10 +192,8 @@ function SolutionState({
     }
   }, [showBeforeState, isReframingStory, isRefiningStory, reducedMotion, totalSections]);
 
-  // Derived: is the panel fully in v1 mode?
   const isFullyBefore = hasToggle && upgradedSections === 0;
 
-  // Get block-level style: dimming + optional background highlight
   const getBlockStyle = (sectionNumber: number) => {
     const transition = 'opacity 0.3s ease-in-out, background-color 0.3s ease-in-out';
 
@@ -808,7 +222,6 @@ function SolutionState({
     return { opacity: 0.4, transition };
   };
 
-  // --- Reframing story: per-section before/after ---
   const reframingBefore = storyInteractions.reframingSummary.before;
   const refiningBefore = storyInteractions.refiningOutput.before;
   const refiningBeforeItems = [
@@ -820,12 +233,12 @@ function SolutionState({
   const isItemUpgraded = (index: number) => upgradedSections >= index + 2;
 
   const renderContent = () => {
-    // For non-toggle stories, render normal after content
     if (!hasToggle) {
       return (
         <>
           <div
             className="border-b-2 border-border px-6 py-6"
+            data-section-id="highlights-section-1"
             style={getBlockStyle(1)}
           >
             <div className="flex items-center gap-3 mb-2">
@@ -845,18 +258,20 @@ function SolutionState({
             <p className="text-body-sm text-primary mt-0">{summary}</p>
           </div>
 
-          {allItems.map((item, index) => (
-            <CollapsibleItem
-              key={`default-${index}`}
-              item={item}
-              sectionIndex={index}
-              isExpanded={expandedIndex === index}
-              onToggle={() =>
-                setExpandedIndex(expandedIndex === index ? null : index)
-              }
-              blockStyle={getBlockStyle(index + 2)}
-            />
-          ))}
+          <div data-section-id="highlights-section-2">
+            {allItems.map((item, index) => (
+              <CollapsibleItem
+                key={`default-${index}`}
+                item={item}
+                sectionIndex={index}
+                isExpanded={expandedIndex === index}
+                onToggle={() =>
+                  setExpandedIndex(expandedIndex === index ? null : index)
+                }
+                blockStyle={getBlockStyle(index + 2)}
+              />
+            ))}
+          </div>
 
           <div className="px-6 py-4 flex items-center gap-2" style={getFooterStyle()}>
             <span className="text-body-sm text-secondary">Is this helpful?</span>
@@ -905,7 +320,6 @@ function SolutionState({
       );
     }
 
-    // --- Toggle stories: per-section streaming ---
     const headerSummaryText = isReframingStory
       ? isHeaderUpgraded ? summary : reframingBefore.summary
       : summary;
@@ -915,8 +329,7 @@ function SolutionState({
 
     return (
       <>
-        {/* Header Section — streams between v1/v2 */}
-        <div className="border-b-2 border-border px-6 py-6">
+        <div className="border-b-2 border-border px-6 py-6" data-section-id="highlights-section-1">
           <div style={{ opacity: isRefiningStory ? 0.3 : 1, transition: 'opacity 0.3s ease-in-out' }}>
             <div className="flex items-center gap-3 mb-2">
               <div
@@ -947,7 +360,7 @@ function SolutionState({
           </div>
         </div>
 
-        {/* Items — each streams independently */}
+        <div data-section-id="highlights-section-2">
         {allItems.map((afterItem, index) => {
           const upgraded = isItemUpgraded(index);
 
@@ -988,7 +401,6 @@ function SolutionState({
             );
           }
 
-          // Refining story: simplified focused view, detail expands in
           const displayItem = upgraded ? afterItem : refiningBeforeItems[index];
           const afterParts = refiningBefore.afterParts?.[index];
           return (
@@ -1009,8 +421,8 @@ function SolutionState({
             </AnimatePresence>
           );
         })}
+        </div>
 
-        {/* Footer */}
         <div
           className="px-6 py-4 flex items-center gap-2"
           style={{
@@ -1066,10 +478,8 @@ export default function HighlightsPanel({
   const hasStartedRef = useRef(false);
 
   useEffect(() => {
-    // Only start timer once intro is complete AND we haven't started yet
     if (isComplete && showLoading && !hasStartedRef.current) {
       hasStartedRef.current = true;
-      // Small delay for fadeInUp to complete, then show loading for 1.5s
       const entranceDelay = reducedMotion ? 0 : 600;
       const loadingDuration = reducedMotion ? 0 : 1500;
 
@@ -1082,6 +492,7 @@ export default function HighlightsPanel({
   }, [isComplete, showLoading, reducedMotion]);
 
   return (
+    <div data-section-id="highlights-panel">
     <AnimatePresence mode="wait">
       {showLoading ? (
         <motion.div
@@ -1109,5 +520,6 @@ export default function HighlightsPanel({
         </motion.div>
       )}
     </AnimatePresence>
+    </div>
   );
 }
